refactor: move and namespace event queue
